requirements.txt
fastapi
httpx
load_dotenv
